+++ finvizfinance/screener/technical.py
@@ -1,5 +1,3 @@
-from finvizfinance.screener.overview import Overview
-
 """
 .. module:: screen.technical
    :synopsis: screen technical table.
@@ -7,8 +5,10 @@ from finvizfinance.screener.overview import Overview
 .. moduleauthor:: Tianning Li <ltianningli@gmail.com>
 """
 
+from finvizfinance.screener.base import Base
 
-class Technical(Overview):
+
+class Technical(Base):
     """Technical inherit from overview module.
     Getting information from the finviz screener technical page.
     """
