+++ finvizfinance/screener/ticker.py
@@ -5,11 +5,16 @@
 .. moduleauthor:: Tianning Li <ltianningli@gmail.com>
 """
 from time import sleep
-from finvizfinance.screener.overview import Overview
-from finvizfinance.util import web_scrap, progress_bar
+from finvizfinance.util import (
+    web_scrap,
+    progress_bar,
+)
+from finvizfinance.constants import order_dict
+
+from finvizfinance.screener.base import Base
 
 
-class Ticker(Overview):
+class Ticker(Base):
     """Financial inherit from overview module.
     Getting information from the finviz screener ticker page.
     """
@@ -25,7 +30,7 @@ class Ticker(Overview):
         return tickers
 
     def screener_view(
-        self, order="ticker", limit=-1, verbose=1, ascend=True, sleep_sec=1
+        self, order="Ticker", limit=-1, verbose=1, ascend=True, sleep_sec=1
     ):
         """Get screener stocks.
 
@@ -38,21 +43,18 @@ class Ticker(Overview):
         Returns:
             tickers(list): get all the tickers as list.
         """
-        url = self.url
-        if order != "ticker":
-            if order not in self.order_dict:
-                order_keys = list(self.order_dict.keys())
-                raise ValueError(
-                    "Invalid order '{}'. Possible order: {}".format(order, order_keys)
-                )
-            url = self.url + "&" + self.order_dict[order]
-        if not ascend:
-            url = url.replace("o=", "o=-")
-        soup = web_scrap(url)
+        if order not in order_dict:
+            order_keys = list(order_dict.keys())
+            raise ValueError(
+                "Invalid order '{}'. Possible order: {}".format(
+                    order, order_keys)
+            )
+        self.request_params['o'] = (
+            "" if ascend else '-') + order_dict[order]
+        soup = web_scrap(self.url, self.request_params)
         page = self._get_page(soup)
         if page == 0:
-            if verbose == 1:
-                print("No ticker found.")
+            print("No ticker found.")
             return None
 
         if limit != -1:
@@ -69,6 +71,7 @@ class Ticker(Overview):
             sleep(sleep_sec)  # Adding sleep
             if verbose == 1:
                 progress_bar(i + 1, page)
-            soup = web_scrap(self.url + "&r={}".format(i * 1000 + 1))
+            self.request_params['r'] = i*1000+1
+            soup = web_scrap(self.url, self.request_params)
             tickers = self._screener_helper(i, page, soup, tickers, limit)
         return tickers
